% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{cash_flow_mat}
\alias{cash_flow_mat}
\title{cash_flow_matrix}
\usage{
cash_flow_mat(tenor, ytm, freq = 2)
}
\arguments{
\item{tenor}{Vector of bond maturities}

\item{ytm}{Vector of yield to maturities}

\item{freq}{Compounding frequency per year (default is 2)}
}
\value{
A matrix with cash flows for different tenors and yield to maturities
}
\description{
Construct cash flow matrix
}
