% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{compute_forwards}
\alias{compute_forwards}
\title{Calculate Forward Rates from Spot Rates}
\usage{
compute_forwards(spot_rates)
}
\arguments{
\item{spot_rates}{A numeric vector of spot rates.}
}
\value{
A numeric vector of forward rates.
}
\description{
Given a vector of spot rates, this function calculates the corresponding forward rates.
}
