% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cont2disc}
\alias{cont2disc}
\title{Convert Interest Rates from Continuous to Discrete (or vice versa)}
\usage{
cont2disc(int, convert = c("discrete", "continuous"))
}
\arguments{
\item{int}{A numeric value representing the interest rate to be converted.}

\item{convert}{A character string specifying the direction of the conversion. Acceptable values are "discrete" and "continuous".}
}
\value{
A numeric value representing the converted interest rate.
}
\description{
This function allows users to convert interest rates between continuous and discrete forms.
}
\examples{
cont2disc(0.05,"discrete")
cont2disc(0.126,"continuous")
}
