% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{discount_pricing}
\alias{discount_pricing}
\title{Discount pricing function with Wilson Function and UFR}
\usage{
discount_pricing(v, tenor, ytm, a, ufr, freq = 2)
}
\arguments{
\item{v}{a vector of tenors for which to calculate the discounted prices.}

\item{tenor}{a vector of the number of years remaining until the bond reaches maturity.}

\item{ytm}{a vector of yields to maturity of observed tenors.}

\item{a}{a parameter for the Wilson function.}

\item{ufr}{the Ultimate Forward Rate.}

\item{freq}{The frequency of coupon payments per year. If freq=0, this indicates a zero-coupon bond.}
}
\value{
a vector of discounted prices for the given tenors.
}
\description{
This function calculates the discounted prices for a given vector of tenors and yields,
taking into account the Ultimate Forward Rate (UFR) and the Wilson function.
}
