% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expSmooth}
\alias{expSmooth}
\title{지수평활화가중평균}
\usage{
expSmooth(vector, lambda = NULL, round = TRUE)
}
\arguments{
\item{vector}{A numeric vector, 평균을 구할 시계열.}

\item{lambda}{A numeric 평활화 승수값.}

\item{round}{Logical 5bp 단위 반올림?}
}
\value{
A numeric.
}
\description{
Compute Exponetially Weighted Moving Average.
}
\examples{
x <- 1:20
expSmooth(x,lambda=0.85)
}
