% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kics-package.R
\docType{package}
\name{kics-package}
\alias{kics}
\alias{kics-package}
\title{kics: STARS functionality for insurance K-ICS stanadrd}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Junho Lee \email{morner@fss.or.kr} (\href{https://orcid.org/0000-0002-4726-210X}{ORCID})

}
\keyword{internal}
