% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nom2eff}
\alias{nom2eff}
\title{Effective interest rate and nominal interest rate converter}
\usage{
nom2eff(int, conversion = 2, convert = c("effective", "nominal"))
}
\arguments{
\item{int}{A decimal number indicating the interest rate. This must be less than or equal to 1.}

\item{conversion}{An integer indicating the number of compounding periods in a year. Default is 2.}

\item{convert}{A string indicating which type of interest rates to convert to. Either "effective" or "nominal". Default is "effective".}
}
\value{
A decimal number representing either the effective interest rate or nominal interest rate.
}
\description{
This function converts either the nominal or effective interest rate to the other one.
}
\examples{
nom2eff(0.12,4,"effective")
nom2eff(0.06,2,"nominal")
}
