% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{wilsonG}
\alias{wilsonG}
\title{Wilson G Function}
\usage{
wilsonG(u, v, a)
}
\arguments{
\item{u}{A vector of numeric values.}

\item{v}{A vector of numeric values.}

\item{a}{A numeric value for parameter a.}
}
\value{
Returns a matrix of numeric values representing the Wilson G function with dimensions length(u) by length(v).
}
\description{
Computes the values of the Wilson G function given u and v.
}
