% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{wilsonH}
\alias{wilsonH}
\title{wilston H}
\usage{
wilsonH(u, v, a)
}
\arguments{
\item{u}{A vector of time values, observed durations to maturity}

\item{v}{A vector of maturity values}

\item{a}{A numeric value, speed of convergence}
}
\value{
Returns a matrix of Wilston's function values
}
\description{
Calculates the Wilson's H function with given parameters
}
