% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{wilsonW}
\alias{wilsonW}
\title{wilston W}
\usage{
wilsonW(u, v, a, ufr)
}
\arguments{
\item{u}{A vector of time values, observed durations to maturity}

\item{v}{A vector of maturity values}

\item{a}{A numeric value, speed of convergence}

\item{ufr}{A numeric value, ultimate forward intensity(continuous rate)}
}
\value{
Returns a matrix of Wilston's function values
}
\description{
Calculates the Wilson's W function with given parameters
}
\examples{
WilsonW(u = c(1, 2, 3), v = c(2, 3, 4), a = 0.05, ufr = 0.03)

}
