% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{ytm2price}
\alias{ytm2price}
\title{Calculate market Bond Price from Yield-to-Maturity}
\usage{
ytm2price(tenor, ytm, freq)
}
\arguments{
\item{tenor}{The number of years remaining until the bond reaches maturity.}

\item{ytm}{The yield-to-maturity of the bond.}

\item{freq}{The frequency of coupon payments per year. If Freq=0, this indicates a zero-coupon bond.}
}
\value{
The price of the bond.
}
\description{
This function calculates the price of a bond with a given tenor, yield-to-maturity, and coupon frequency.
}
