% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithwilson.R
\name{ytm2price}
\alias{ytm2price}
\title{Yield to maturity to market price}
\usage{
ytm2price(tenor, ytm, freq)
}
\arguments{
\item{tenor}{length of time until maturity of the bond, in years.}

\item{ytm}{the yield to maturity of the bond.}

\item{freq}{the number of coupon payments per year.}
}
\value{
The price of the bond.
}
\description{
Computes the price of a bond given its tenor, yield to maturity, and coupon frequency.
}
