% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{YTM2Price}
\alias{YTM2Price}
\title{Calculate market Bond Price from Yield-to-Maturity}
\usage{
YTM2Price(Tenor, YTM, Freq)
}
\arguments{
\item{Tenor}{The number of years remaining until the bond reaches maturity.}

\item{YTM}{The yield-to-maturity of the bond.}

\item{Freq}{The frequency of coupon payments per year. If Freq=0, this indicates a zero-coupon bond.}
}
\value{
The price of the bond.
}
\description{
This function calculates the price of a bond with a given tenor, yield-to-maturity, and coupon frequency.
}
\examples{
YTM2Price(5, 0.03, 2)
YTM2Price(10, 0.05, 4)

}
